# server/swagger.yaml
openapi: 3.0.0
info:
  title: Grihastha API Documentation
  description: Complete REST API documentation for the Grihastha Property Rental Platform. Exposes endpoints for Users, Hosts, and Admins.
  version: 1.0.0
servers:
  - url: http://localhost:5001
    description: Development Server
paths:
...
  /v1/listings:
    post:
      summary: Create a new blank listing (DRAFT status)
      tags:
        - Property Listings
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Listing created successfully with DRAFT status
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  host_id:
                    type: string
                    format: uuid
                  category:
                    type: string
                    nullable: true
                  address:
                    type: object
                    nullable: true
                  floor_plan:
                    type: object
                    nullable: true
                  amenities:
                    type: array
                    nullable: true
                  photos:
                    type: array
                    nullable: true
                  status:
                    type: string
                    enum: [DRAFT, PUBLISHED]
                  created_at:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized - Host JWT required
...
  /v1/bookings:
    post:
      summary: Create booking (instant or request-to-book)
      tags: [Bookings]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listing_id, check_in, check_out, booking_type]
              properties:
                listing_id:
                  type: string
                  format: uuid
                check_in:
                  type: string
                  format: date
                  example: "2025-07-10"
                check_out:
                  type: string
                  format: date
                  example: "2025-07-14"
                guests:
                  type: object
                  properties:
                    adults:
                      type: integer
                      example: 2
                    children:
                      type: integer
                      example: 1
                    infants:
                      type: integer
                      example: 0
                    pets:
                      type: integer
                      example: 0
                booking_type:
                  type: string
                  enum: [instant, request]
                  example: "instant"
                special_requests:
                  type: string
                  example: "Early check-in if possible."
      responses:
        '201':
          description: Booking created with price breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking_id:
                    type: string
                    format: uuid
                  nights:
                    type: integer
                  status:
                    type: string
                    enum: [pending, confirmed]
                  price_breakdown:
                    type: object
                    properties:
                      base_price:
                        type: integer
                      cleaning_fee:
                        type: integer
                      platform_service_fee:
                        type: integer
                      tax:
                        type: integer
                      total:
                        type: integer
                  payment_status:
                    type: string
                    enum: [unpaid, paid]
        '400':
          description: Validation error (dates, own listing, etc.)
        '409':
          description: Dates already booked
    get:
      summary: Guest booking history
      tags: [Bookings]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled, rejected]
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of guest bookings
...
  /v1/wishlists:
    post:
      summary: Create a new wishlist
      tags: [Wishlists]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "Dream Nepal Stays"
      responses:
        '201':
          description: Wishlist created
    get:
      summary: Get all user wishlists
      tags: [Wishlists]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of wishlists with cover image and count
...
  /v1/reviews:
    post:
      summary: Guest submits review after checkout
      tags: [Reviews]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [booking_id, overall_rating]
              properties:
                booking_id:
                  type: string
                  format: uuid
                overall_rating:
                  type: number
                  minimum: 1
                  maximum: 5
                  example: 4.5
                cleanliness:
                  type: number
                accuracy:
                  type: number
                communication:
                  type: number
                location:
                  type: number
                check_in:
                  type: number
                value:
                  type: number
                comment:
                  type: string
                  example: "Great stay! Beautiful views."
      responses:
        '201':
          description: Review created
        '400':
          description: Already reviewed or booking not completed
...
  /v1/conversations:
    post:
      summary: Start conversation with a host
      tags: [Messaging]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host_id]
              properties:
                host_id:
                  type: string
                  format: uuid
                listing_id:
                  type: string
                  format: uuid
                message:
                  type: string
                  example: "Hi! Is your place available?"
      responses:
        '200':
          description: Conversation created
    get:
      summary: List all conversations
      tags: [Messaging]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of conversations
...
  /v1/notifications:
    get:
      summary: List notifications
      tags: [Notifications]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: unread_only
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of notifications
...
  /v1/admin/users:
    get:
      summary: List all platform users
      tags: [Admin Management]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [user, host, admin]
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated user list
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token required for protected endpoints